backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
python-dotenv==1.0.0